+++ tests/testthat/test-derive_expected_records.R
@@ -28,12 +28,12 @@ test_that("derive_expected_records Test 1: missing values in `by_vars`", {
       mutate(DTYPE = "DERIVED")
   )
 
-  actual_output <- derive_expected_records(
+  actual_output <- suppressMessages(derive_expected_records(
     dataset = input,
     dataset_ref = expected_obsv,
     by_vars = exprs(USUBJID),
     set_values_to = exprs(DTYPE = "DERIVED")
-  )
+  ))
 
   expect_dfs_equal(
     base = expected_output,
@@ -69,12 +69,12 @@ test_that("derive_expected_records Test 2: `by_vars` = NULL", {
       mutate(DTYPE = "DERIVED")
   )
 
-  actual_output <- derive_expected_records(
+  actual_output <- suppressMessages(derive_expected_records(
     dataset = input,
     dataset_ref = expected_obsv,
     by_vars = NULL,
     set_values_to = exprs(DTYPE = "DERIVED")
-  )
+  ))
 
   expect_dfs_equal(
     base = expected_output,
@@ -112,12 +112,12 @@ test_that("derive_expected_records Test 3: visit variables are parameter indepen
       mutate(DTYPE = "DERIVED")
   )
 
-  actual_output <- derive_expected_records(
+  actual_output <- suppressMessages(derive_expected_records(
     dataset = input,
     dataset_ref = expected_obsv,
     by_vars = exprs(USUBJID, PARAMCD),
     set_values_to = exprs(DTYPE = "DERIVED")
-  )
+  ))
 
   expect_dfs_equal(
     base = expected_output,
@@ -155,16 +155,44 @@ test_that("derive_expected_records Test 4: visit variables are parameter depende
       mutate(DTYPE = "DERIVED")
   )
 
-  actual_output <- derive_expected_records(
+  actual_output <- suppressMessages(derive_expected_records(
     dataset = input,
     dataset_ref = expected_obsv,
     by_vars = exprs(USUBJID),
     set_values_to = exprs(DTYPE = "DERIVED")
-  )
+  ))
 
   expect_dfs_equal(
     base = expected_output,
     compare = actual_output,
     keys = c("USUBJID", "PARAMCD", "AVISITN", "AVISIT", "DTYPE")
+  )
+})
+
+
+## Test 5: message reports matching variables and records added ----
+test_that("derive_expected_records Test 5: message reports matching variables and records added", {
+  input <- tibble::tribble(
+    ~USUBJID, ~PARAMCD, ~AVISITN, ~AVISIT, ~AVAL,
+    "1", "a", 1, "WEEK 1", 10,
+    "1", "b", 2, "WEEK 2", 11
+  )
+
+  expected_obsv <- tibble::tribble(
+    ~PARAMCD, ~AVISITN, ~AVISIT,
+    "a", 1, "WEEK 1",
+    "a", 2, "WEEK 2",
+    "b", 1, "WEEK 1",
+    "b", 2, "WEEK 2"
+  )
+
+  expect_message(
+    derive_expected_records(
+      dataset = input,
+      dataset_ref = expected_obsv,
+      by_vars = NULL,
+      set_values_to = exprs(DTYPE = "DERIVED")
+    ),
+    "Expected observations identified"
   )
 })